task 6 implement complete
may need to revisit task 4 and 5 (flip and split) implementations, one or both is connecting half_edges wrong or grabbing the wrong neighbors. 
task 6 seems to be nearly working but new hiccups in 4 and 5 have appeared as 6 pushes their usage to the max compared to my individual testing with the provided key toggles. Circle render seems okay but sharper edged shapes (cube, quad, etc) are not being maintained and loose quite a bit of integrity T^T
onward !!